logged in and owner check

+++ apps/account/views.py
@@ -6,6 +6,7 @@ from apps.account.services.account_service import AccountService
 from django.contrib.auth.models import AnonymousUser
 from apps.account.forms import AccountCreateForm, AccountTransferForm
 from utils.exceptions import AuthException
+from utils.permissions import logged_in
 
 
 class AccountDetailView(View):
@@ -13,14 +14,18 @@ class AccountDetailView(View):
     service = AccountService()
     account_transfer_form = AccountTransferForm
 
+    @logged_in
     def get(self, request, pk):
         context = {
             'account_transfer_form': self.account_transfer_form(request.user)
         }
         account = self.service.retrieve_account_by_pk(pk=pk)
-        if account:
-            context['account'] = self.service.get_account_context(account)
-        return render(request, template_name=self.template_name, context=context)
+        if account.owner == request.user:
+            if account:
+                context['account'] = self.service.get_account_context(account)
+            return render(request, template_name=self.template_name, context=context)
+        else:
+            raise AuthException()
 
 
 class AccountListView(View):
@@ -28,15 +33,13 @@ class AccountListView(View):
     service = AccountService()
     account_create_form = AccountCreateForm
 
+    @logged_in
     def get(self, request):
         context = {"account_create_form": self.account_create_form}
 
-        if request.user is not AnonymousUser:
-            user = request.user
-            accounts = self.service.retrieve_user_accounts(user=user)
-            context['accounts'] = self.service.get_account_context(accounts, many=True)
-        else:
-            raise AuthException()
+        user = request.user
+        accounts = self.service.retrieve_user_accounts(user=user)
+        context['accounts'] = self.service.get_account_context(accounts, many=True)
 
         return render(request, template_name='account/account_list.html', context=context)
 
@@ -44,6 +47,7 @@ class AccountListView(View):
 class AccountCreateView(View):
     service = AccountService()
 
+    @logged_in
     def post(self, request):
         form = AccountCreateForm(request.POST)
         if form.is_valid():
@@ -54,16 +58,22 @@ class AccountCreateView(View):
 class AccountDeleteView(View):
     service = AccountService()
 
+    @logged_in
     def delete(self, request, pk):
-        if self.service.delete_account(pk):
-            return redirect("account_list")
+        account = self.service.retrieve_account_by_pk(pk=pk)
+        if account.owner == request.user:
+            if self.service.delete_account(pk):
+                return redirect("account_list")
+            else:
+                return redirect("account_detail", pk)
         else:
-            return redirect("account_detail", pk)
+            raise AuthException()
 
 
 class AccountTransferView(View):
     service = AccountService()
 
+    @logged_in
     def post(self, request, pk):
         form = AccountTransferForm(request.user, request.POST)
         context = {"account_transfer_form": form}
@@ -74,7 +84,11 @@ class AccountTransferView(View):
             source = pk
             own_account = form.cleaned_data["own_accounts"]
             if own_account != "--":
-                self.service.execute_account_transaction(str(source), str(own_account), amount)
+                account = self.service.retrieve_account_by_pk(pk=pk)
+                if account.owner == request.user:
+                    self.service.execute_account_transaction(str(source), str(own_account), amount)
+                else:
+                    raise AuthException()
             elif destination is not None:
                 self.service.execute_account_transaction(str(source), str(destination), amount)
             else:
